Dockerized/app/requirements.txt
streamlit
mysql-connector-python
pillow
tensorflow
